move eleventy-test-instance declaration to tests root
