+++ jsconfig.json
@@ -3,8 +3,5 @@
 		"module": "NodeNext",
 		"moduleResolution": "NodeNext",
 		"baseUrl": "./",
-		"paths": {
-			"#11ty-test": ["./tests/_utils/eleventy-test.js"]
-		}
 	}
 }
